fix: Campaign-2 Phase B/C defect-sweep hardening and contract annotations (passes 157-193)
Source changes from the adversarial CTF/KISS audit plus contract/port-layer
work accumulated since pass 156:

- arm/spectral_debug_embedded_arm.c: fix unsigned-EWMA underflow in the
  DWT/ITM timing and SD-latency rolling averages (signed-delta arithmetic
  at all three sites) — the one reachable defect found in the sweep (187).
- core/spectral_osc_formulas.h: clamp asin oscillator to [-1,1] before
  asinf so an out-of-domain argument cannot inject NaN into the mix.
- synth/math/spectral_q15.h, core/spectral_synth_internal.c: finite/range
  guards before float->int and double->size_t conversions.
- core/spectral_segment_parser.{c,h}, spectral_config.h: segment/stretch
  validation and bounds.
- sim/spectral_synth_simulation.c: rework the simulation into a perf/
  resource model over the real arm32 core (net code reduction).
- core/spectral_contracts.h + analysis/synth/core/port call sites: COLA
  and invariant contract annotations; cuda/metal/gpu_tile parity touches.

Triad green: five production targets build clean, ctest 4/4.

+++ spectral_engine/core/port/embedded/spectral_out_kernels.c
@@ -98,7 +98,7 @@ void spectral_mono_to_stereo_q15(const q15_t* mono, q15_t* stereo, size_t num_fr
     if (!mono || !stereo || num_frames == 0 || num_frames > SIZE_MAX / 2u) return;
 
 #if SPECTRAL_ARM_M7 && defined(__ARM_FEATURE_DSP)
-    size_t pairs = num_frames & ~1U;
+    size_t pairs = num_frames & ~(size_t)1;
     size_t i = 0;
     for (; i < pairs; i += 2) {
         size_t stereo_i = i * 2u;
